gui: moved functions to other file
these are SLOTS for buttons and lineEdits in Sections
